It moves! Camera is janky
so i unchecked the box collider attached to the player. He now moves. the if statement was always true because the player WAS a box collider. if you add an empty object with box collider to the game he can't pass over it. There's a test object i just kinda stuck in there. The camera is super janky though